Add more info logging per request of DMS
